requirements.txt
appdirs==1.4.4
asgiref==3.4.1
black==21.7b0
certifi==2021.5.30
cffi==1.14.6
charset-normalizer==2.0.6
click==8.0.1
cloudinary==1.26.0
colorama==0.4.4
coreapi==2.3.3
coreschema==0.0.4
cryptography==35.0.0
defusedxml==0.7.1
Deprecated==1.2.13
dj-database-url==0.5.0
dj-rest-auth==2.1.11
Django==3.2.8
django-allauth==0.45.0
django-cloudinary-storage==0.3.0
django-cors-headers==3.10.0
django-environ==0.8.1
django-filter==21.1
django-heroku==0.3.1
django-oauth-toolkit==1.5.0
django-quill-editor==0.1.22
django-rest-framework==0.1.0
djangorestframework==3.12.4
djangorestframework-simplejwt==4.6.0
drf-social-oauth2==1.0.9
drf-yasg==1.20.0
gunicorn==20.1.0
idna==3.2
inflection==0.5.1
itypes==1.2.0
Jinja2==3.0.2
jwcrypto==1.0
MarkupSafe==2.0.1
mypy-extensions==0.4.3
ndg-httpsclient==0.5.1
oauthlib==3.1.1
packaging==21.0
pathspec==0.9.0
Pillow==8.3.2
psycopg2==2.9.1
pusher==3.1.0
pyasn1==0.4.8
pycparser==2.20
PyJWT==2.2.0
PyNaCl==1.4.0
pyOpenSSL==21.0.0
pyparsing==2.4.7
python3-openid==3.2.0
pytz==2021.3
regex==2021.8.3
requests==2.26.0
requests-oauthlib==1.3.0
ruamel.yaml==0.17.16
ruamel.yaml.clib==0.2.6
shortuuid==1.0.1
six==1.16.0
social-auth-app-django==5.0.0
social-auth-core==4.1.0
sqlparse==0.4.2
tomli==1.2.1
uritemplate==3.0.1
urllib3==1.26.7
whitenoise==5.3.0
wrapt==1.13.1
